Make lidar and GPS sensors more realistic
GPS now only provides new data once per second, and it provides no
heading information. This is all that our current hardware GPS device
provides.

Lidar now includes hits for landmarks. This is assuming that landmarks
are thin vertical obstacles (which is true for URC). It is also assuming
that they are not SO thin that our lidar can't reliably see them. I am
not sure how true that second assumption is.