Change 2D_slam example to handle no heading GPS information
Introduces a new class FriendlyGraph that's much more intuitive for
adding different kinds of factors.

Adds GPS measurements to the 2D_slam example.

The starting pose is now itself an optimization variable, rather than
being defined to be (0,0,0). This is so that all variables are now in
the GPS frame.

Currently 1D_slam will not compile; I am sorry.